refactor: change testing file
