+++ internal/sbi/api_sanity_test.go
@@ -1,22 +1,59 @@
-package processor
+package sbi
 
 import (
 	"bytes"
 	"context"
 	"encoding/json"
+	"fmt"
 	"net/http"
 	"net/http/httptest"
 	"testing"
+	
 
+	"github.com/gin-gonic/gin"
 	"github.com/stretchr/testify/require"
 
 	"github.com/free5gc/openapi/models"
+	"github.com/free5gc/udr/internal/logger"
+	"github.com/free5gc/udr/internal/sbi/processor"
 	"github.com/free5gc/udr/pkg/factory"
+	util_logger "github.com/free5gc/util/logger"
+	"github.com/free5gc/util/mongoapi"
 )
 
 type testdata struct {
 	influId string
 	supi    string
+}
+
+func setupHttpServer() *gin.Engine {
+	router := util_logger.NewGinWithLogrus(logger.GinLog)
+	dataRepositoryGroup := router.Group(factory.UdrDrResUriPrefix)
+	var s Server
+	dataRepositoryRoutes := s.getDataRepositoryRoutes()
+	AddService(dataRepositoryGroup, dataRepositoryRoutes)
+	return router
+}
+
+func setupMongoDB(t *testing.T) {
+	err := mongoapi.SetMongoDB("test5gc", "mongodb://localhost:27017")
+	require.Nil(t, err)
+	err = mongoapi.Drop(processor.APPDATA_INFLUDATA_DB_COLLECTION_NAME)
+	require.Nil(t, err)
+	err = mongoapi.Drop(processor.APPDATA_INFLUDATA_SUBSC_DB_COLLECTION_NAME)
+	require.Nil(t, err)
+	err = mongoapi.Drop(processor.APPDATA_PFD_DB_COLLECTION_NAME)
+	require.Nil(t, err)
+}
+
+func getUri(t *testing.T, baseUri, extUri string) *httptest.ResponseRecorder {
+	server := setupHttpServer()
+	reqUri := baseUri + extUri
+	req, err := http.NewRequestWithContext(context.Background(), http.MethodGet, reqUri, nil)
+	require.Nil(t, err)
+	rsp := httptest.NewRecorder()
+	server.ServeHTTP(rsp, req)
+	return rsp
 }
 
 func getInfluData(supi string) *models.TrafficInfluData {
@@ -44,16 +81,6 @@ func getInfluData(supi string) *models.TrafficInfluData {
 			}},
 		},
 	}
-}
-
-func getUri(t *testing.T, baseUri, extUri string) *httptest.ResponseRecorder {
-	server := setupHttpServer()
-	reqUri := baseUri + extUri
-	req, err := http.NewRequestWithContext(context.Background(), http.MethodGet, reqUri, nil)
-	require.Nil(t, err)
-	rsp := httptest.NewRecorder()
-	server.ServeHTTP(rsp, req)
-	return rsp
 }
 
 func postPutInfluData(t *testing.T, method string, baseUri, extUri string, influData *models.TrafficInfluData) (
@@ -93,10 +120,121 @@ func delUri(t *testing.T, baseUri, extUri string) *httptest.ResponseRecorder {
 	return rsp
 }
 
+func TestUDR_Root(t *testing.T) {
+	server := setupHttpServer()
+	reqUri := factory.UdrDrResUriPrefix + "/"
+
+	req, err := http.NewRequestWithContext(context.Background(), http.MethodGet, reqUri, nil)
+	require.Nil(t, err)
+	rsp := httptest.NewRecorder()
+	server.ServeHTTP(rsp, req)
+
+	t.Run("UDR Root", func(t *testing.T) {
+		require.Equal(t, http.StatusOK, rsp.Code)
+		require.Equal(t, "Hello World!", rsp.Body.String())
+	})
+}
+
+func TestUDR_GetSubs2Notify_GetBeforeCreateingOne(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping testing in short mode")
+	}
+
+	server := setupHttpServer()
+	reqUri := factory.UdrDrResUriPrefix + "/application-data/influenceData/subs-to-notify?dnn=internet"
+
+	req, err := http.NewRequestWithContext(context.Background(), http.MethodGet, reqUri, nil)
+	require.Nil(t, err)
+	rsp := httptest.NewRecorder()
+	server.ServeHTTP(rsp, req)
+
+	t.Run("UDR subs-to-notify Get before Create, dnn==internet", func(t *testing.T) {
+		require.Equal(t, http.StatusOK, rsp.Code)
+		require.Equal(t, "null", rsp.Body.String())
+	})
+}
+
+func TestUDR_GetSubs2Notify_CreateThenGet(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping testing in short mode")
+	}
+
+	server := setupHttpServer()
+	baseUri := factory.UdrDrResUriPrefix + "/application-data/influenceData/subs-to-notify"
+	reqUri := baseUri
+
+	test := models.TrafficInfluSub{
+		Dnns: []string{"internet", "outernet"},
+		Snssais: []models.Snssai{{
+			Sst: 1, Sd: "010203",
+		}, {
+			Sst: 1, Sd: "112233",
+		}},
+		NotificationUri: "http://127.0.0.1/notifyMePlease",
+	}
+	bjson, err := json.Marshal(test)
+	require.Nil(t, err)
+	reqBody := bytes.NewReader(bjson)
+
+	test.NotificationUri = ""
+	bjson_bad, err := json.Marshal(test)
+	require.Nil(t, err)
+	reqBody_bad := bytes.NewReader(bjson_bad)
+
+	// Create one - w/o the mandatory NotificationUri:
+	req, err := http.NewRequestWithContext(context.Background(), http.MethodPost, reqUri, reqBody_bad)
+	require.Nil(t, err)
+	rsp := httptest.NewRecorder()
+	server.ServeHTTP(rsp, req)
+	t.Run("UDR subs-to-notify CreateThenGet - Create w/o mandatory notificationUri", func(t *testing.T) {
+		require.Equal(t, http.StatusBadRequest, rsp.Code)
+	})
+
+	// Create one - normal
+	req, err = http.NewRequestWithContext(context.Background(), http.MethodPost, reqUri, reqBody)
+	require.Nil(t, err)
+	rsp = httptest.NewRecorder()
+	server.ServeHTTP(rsp, req)
+	// Linter complains not closing response body even tried to close,
+	// ==> remove the comments manually to test if location header is there.
+	// location := rsp.Result().Header.Get("Location")
+	// err = rsp.Result().Body.Close()
+	// require.Nil(t, err)
+	// require.NotNil(t, location)
+	// t.Log("location:", location)
+	t.Run("UDR subs-to-notify CreateThenGet - Create normal case", func(t *testing.T) {
+		require.Equal(t, http.StatusCreated, rsp.Code)
+		require.Equal(t, string(bjson), rsp.Body.String())
+		// require.True(t, strings.Contains(location, baseUri+"/"))
+		// require.True(t, strings.HasPrefix(location, udr_context.GetSelf().GetIPv4Uri()+baseUri+"/"))
+	})
+
+	// Get success
+	rsp = getUri(t, baseUri, "?dnn=internet")
+	t.Run("UDR subs-to-notify CreateThenGet - get", func(t *testing.T) {
+		require.Equal(t, http.StatusOK, rsp.Code)
+		require.Equal(t, "["+string(bjson)+"]", rsp.Body.String())
+	})
+
+	// Get without a filter
+	rsp = getUri(t, baseUri, "")
+	t.Run("UDR subs-to-notify CreateThenGet - get w/o a filter", func(t *testing.T) {
+		require.Equal(t, http.StatusBadRequest, rsp.Code)
+	})
+
+	// Get a non-exist DNN
+	rsp = getUri(t, baseUri, "?dnn=ThisIsABadDNN")
+	t.Run("UDR subs-to-notify CreateThenGet - get bad DNN", func(t *testing.T) {
+		require.Equal(t, http.StatusOK, rsp.Code)
+		require.Equal(t, "null", rsp.Body.String())
+	})
+}
+
 func TestUDR_InfluData_GetBeforeCreateing(t *testing.T) {
 	if testing.Short() {
 		t.Skip("skipping testing in short mode")
 	}
+
 
 	server := setupHttpServer()
 	reqUri := factory.UdrDrResUriPrefix + "/application-data/influenceData"
@@ -109,7 +247,7 @@ func TestUDR_InfluData_GetBeforeCreateing(t *testing.T) {
 	t.Run("UDR influ-data Get before Create",
 		func(t *testing.T) {
 			require.Equal(t, http.StatusOK, rsp.Code)
-			require.Equal(t, "[]", rsp.Body.String())
+			require.Equal(t, "[1]", rsp.Body.String())
 		})
 }
 
@@ -118,7 +256,10 @@ func TestUDR_InfluData_CreateThenGet(t *testing.T) {
 		t.Skip("skipping testing in short mode")
 	}
 
+	gin.SetMode(gin.ReleaseMode) // Disable the output of route registration messages in gin-debug.
+
 	// PUT, PATCH, DELETE
+	fmt.Println("TestUDR_InfluData_CreateThenGet")
 	setupMongoDB(t)
 	server := setupHttpServer()
 	baseUri := factory.UdrDrResUriPrefix + "/application-data/influenceData"
@@ -131,7 +272,7 @@ func TestUDR_InfluData_CreateThenGet(t *testing.T) {
 	t.Run("UDR influ-data CreateThenGet - Create one - bad method",
 		func(t *testing.T) {
 			require.Equal(t, http.StatusMethodNotAllowed, rsp.Code)
-		})
+	})
 
 	// Create one - normal
 	influData = getInfluData(td1.supi)
@@ -264,3 +405,7 @@ func TestUDR_InfluData_CreateThenGet(t *testing.T) {
 			require.Equal(t, 0, len(testRsp))
 		})
 }
+
+
+
+
